requirements.txt
beautifulsoup4
chromadb
google-generativeai
langchain
langchain-google-genai
pandas
python-dotenv
ragas>=0.1.7
requests
sentence-transformers
streamlit
tqdm
rouge-score
